backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Pillow==10.1.0
pandas==2.1.3
openpyxl==3.1.2
gunicorn==21.2.0
reportlab==4.0.7
